Fixes/QoL stuff
When an important item is given, the return pipe is removed and re-added to the important items, effectively sorting it to the top of the inventory.

Prevent giving back control to the player upon warping via the return pipe.

Prevent setting starting stats if AP is not enabled.

Several other small optimizations.